fix: adjust heart icon position
