+++ src/app/Components/ArtworkGrids/MasonryInfiniteScrollArtworkGrid.tsx
@@ -4,9 +4,6 @@ import { MasonryFlashList, MasonryFlashListProps } from "@shopify/flash-list"
 import { PriceOfferMessage } from "app/Components/ArtworkGrids/ArtworkGridItem"
 import { MasonryArtworkGridItem } from "app/Components/ArtworkGrids/MasonryArtworkGridItem"
 import { PartnerOffer } from "app/Scenes/Activity/components/PartnerOfferCreatedNotification"
-import { getNewSaveAndFollowOnArtworkCardExperimentVariant } from "app/Scenes/Artwork/utils/getNewSaveAndFollowOnArtworkCardExperimentVariant"
-import { useExperimentVariant } from "app/utils/experiments/hooks"
-import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
 import {
   ESTIMATED_MASONRY_ITEM_SIZE,
   MASONRY_LIST_PAGE_SIZE,
@@ -74,18 +71,6 @@ export const MasonryInfiniteScrollArtworkGrid: React.FC<MasonryInfiniteScrollArt
   const space = useSpace()
   const shouldDisplaySpinner = !!artworks.length && !!isLoading && !!hasMore
   const shouldDisplayHeader = !!artworks.length && ListHeaderComponent !== undefined
-  const enableNewSaveAndFollowOnArtworkCard = useFeatureFlag(
-    "AREnableNewSaveAndFollowOnArtworkCard"
-  )
-
-  const newSaveAndFollowOnArtworkCardExperiment = useExperimentVariant(
-    "onyx_artwork-card-save-and-follow-cta-redesign"
-  )
-
-  const { enableShowOldSaveCTA } = getNewSaveAndFollowOnArtworkCardExperimentVariant(
-    newSaveAndFollowOnArtworkCardExperiment.enabled,
-    newSaveAndFollowOnArtworkCardExperiment.variant
-  )
 
   const onEndReached = useCallback(() => {
     if (!!hasMore && !isLoading && !!loadMore) {
@@ -117,13 +102,6 @@ export const MasonryInfiniteScrollArtworkGrid: React.FC<MasonryInfiniteScrollArt
         onPress={onPress}
         hideSaleInfo={hideSaleInfo}
         hideSaveIcon={hideSaveIcon}
-        positionCTAs={
-          enableShowOldSaveCTA || !enableNewSaveAndFollowOnArtworkCard
-            ? "row"
-            : rest.numColumns !== 1
-              ? "column"
-              : "row"
-        }
       />
     ),
     [rest.numColumns]
